skills/wechat-article-writer/python/requirements.txt
matplotlib>=3.5.0
Pillow>=9.0.0
pyyaml>=6.0